switch to unsigned long

+++ chest.cpp
@@ -5,7 +5,7 @@ void Chest::setTexture(const sf::Texture& texture)
     sprite.setTexture(texture);
 }
 
-void Chest::set_position(int x, int y)
+void Chest::set_position(unsigned long x, unsigned long y)
 {
     sprite.setPosition((32 * x) - 32, (32 * y) - 32);
 }
